WSInputs.java - Changed name of DRV_Button_2 to LIFT_BUTTON
WSOutputs.java - Added the outputs RIGHTLOWPIS, LEFTLOWPIS,
RIGHTHIGHPIS, LEFTHIGHPIS, RIGHTHOOK, LEFTHOOK, RIGHTWINCH, LEFTWINCH

WSSubsystems.java - Added "CLIMBER" because it was not there.

Climber.java - Half finished lift/climb code. Added a boolean, checks if
a button is pressed, after checking it sends the value of the boolean to
the pistons.